fix(PrintedBooks): delete ParisSydrac1486
The deleted item conflates publication details of GW M42011 and STC
3186, the second of which is elsewhere recorded in PrintedBooks.xml.

The identical xml:id is employed in Bibliography.xml, where it is
applied correctly to GW M42011. (The duplication is noted in #9.)

The xml:id is referenced once within Records.xml, but not as a witness
to any DIMEV item. As such, the referenced publication should be
recorded within Bibliography.xml not PrintedBooks.xml.