fix(Bibliography): correct variant URLs
The file Bibliography.xml records URLs for on-line facsimiles in the
titlestmt block and again in the pubstmt block. As preparation for
moving the URLs to Manuscripts.xml, the script update-Manuscripts.py
cross-checks the duplicate URLs against one another. In a few cases, the
URLs differed. This commit establishes identity.